requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
coverage==4.5.3
importlib-metadata==0.18
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
mypy==0.711
mypy-extensions==0.4.1
packaging==19.0
pluggy==0.12.0
py==1.8.0
pylint==2.3.1
pyparsing==2.4.0
pytest==4.6.3
pytest-cov==2.7.1
python-dotenv==0.10.3
six==1.12.0
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.1
